cam-watch/requirements.txt
streamlink==6.10.0
google-cloud-storage==2.18.2
opencv_contrib_python==4.10.0.84
